+++ src/databricks/labs/ucx/mixins/fixtures.py
@@ -31,6 +31,7 @@ from databricks.sdk.service.catalog import (
     TableInfo,
     TableType,
 )
+from databricks.sdk.service.iam import WorkspacePermission
 from databricks.sdk.service.serving import (
     EndpointCoreConfigInput,
     ServedModelInput,
@@ -46,6 +47,7 @@ from databricks.sdk.service.sql import (
 )
 from databricks.sdk.service.workspace import ImportFormat
 
+from databricks.labs.ucx.workspace_access.groups import MigratedGroup
 from databricks.labs.ucx.workspace_access.manager import PermissionManager
 
 # this file will get to databricks-labs-pytester project and be maintained/refactored there
@@ -148,11 +150,6 @@ def ws(product_info, debug_env) -> WorkspaceClient:
 
 
 @pytest.fixture
-def permission_manager(ws, sql_backend, inventory_schema) -> PermissionManager:
-    return PermissionManager(ws, sql_backend, inventory_schema, [])
-
-
-@pytest.fixture
 def acc(product_info, debug_env) -> AccountClient:
     # Use variables from Unified Auth
     # See https://databricks-sdk-py.readthedocs.io/en/latest/authentication.html
@@ -164,6 +161,11 @@ def acc(product_info, debug_env) -> AccountClient:
         product=product_name,
         product_version=product_version,
     )
+
+
+@pytest.fixture
+def permission_manager(ws, sql_backend, inventory_schema) -> PermissionManager:
+    return PermissionManager.factory(ws, sql_backend, inventory_schema)
 
 
 def _permissions_mapping():
@@ -645,6 +647,20 @@ def make_group(ws, make_random):
 @pytest.fixture
 def make_acc_group(acc, make_random):
     yield from _make_group("account group", acc.config, acc.groups, make_random)
+
+
+@pytest.fixture
+def make_migrated_group(acc, ws, make_group, make_acc_group):
+    """Create a pair of groups in workspace and account. Assign account group to workspace."""
+
+    def inner():
+        ws_group = make_group()
+        acc_group = make_acc_group()
+        acc.workspace_assignment.update(ws.get_workspace_id(), acc_group.id, [WorkspacePermission.USER])
+        # need to return both, as acc_group.id is not in MigratedGroup dataclass
+        return MigratedGroup.partial_info(ws_group, acc_group), acc_group
+
+    return inner
 
 
 @pytest.fixture
